+++ app/components/AboutHelp.jsx
@@ -3,7 +3,7 @@
 // 2016-18 Claus Zinn, University of Tuebingen
 // 
 // File: AboutHelp.jsx
-// Time-stamp: <2018-09-27 13:33:56 (zinn)>
+// Time-stamp: <2018-11-09 09:26:43 (zinn)>
 // -------------------------------------------
 
 import React from 'react';
@@ -45,25 +45,34 @@ class AboutHelpText extends React.Component {
 	  
             <h2>About</h2>
             <p>
-              The Language Resource Switchboard (LRS) is being developed 
-              within the <a href="http://www.clarin.eu/">CLARIN-PLUS</a> project
-              as a means to link linguistic resources with the tools that can process them.
-	      <em>Tool developers</em>, please contact us in case you want your tool registered with
-              the switchboard!
+              The Language Resource Switchboard (LRS) has been developed 
+              within the <a href="http://www.clarin.eu/">CLARIN-PLUS</a> project. It helps 
+	      users to find and start tools that can process their research data.
             </p>
             
             <h3>Documentation</h3>
             <ul>
-	      <li>Users, please consult "Help" navigation bar.</li>
+	      <li>Users, please consult "Help" navigation bar. There is also a FAQ, see page footer.</li>
 	      <li>Developers, please consult "For Developers" in the footer of the page.</li>	      
-              <li>Frequently Asked Questions, please ask questions now, so we can add them to the FAQ!</li>
             </ul>
 	    
             <h3>Publications</h3>
             <ul>
-              <li>Claus Zinn. <a href="https://www.clarin.eu/sites/default/files/zinn-CLARIN2016_paper_26.pdf">The CLARIN Language Resource Switchboard. </a>CLARIN 2016 Annual Conference, Aix-en-Provence, France, 2016</li>
-              <li>Claus Zinn, Marie Hinrichs, Emanuel Dima, Dieter van Uytvanck. <a href="https://office.clarin.eu/v/CE-2015-0684-LR_switchboard_spec.pdf">The Switchboard specification </a>(Milestone 2.2 of the CLARIN-PLUS project)</li>
-              <li>Claus Zinn. <a href="https://office.clarin.eu/v/CE-2016-0881-CLARINPLUS-D2_5.pdf">D2.5 LR Switchboard (software)</a> Deliverable in the CLARIN-PLUS project.</li>	    
+              <li>Claus Zinn. <a href="http://www.ep.liu.se/ecp/147/004/ecp17147004.pdf">A Bridge
+              from EUDAT's B2DROP cloud service to CLARIN's Language Resource
+              Switchboard </a>Selected papers from the CLARIN Annual Conference 2017, Budapest,
+              18-20 September 2017, Linköping University Electronic Press vol. 147, pages 36-45,
+              2018.</li>
+	      <li>Claus
+	      Zinn. <a href="https://www.clarin.eu/sites/default/files/zinn-CLARIN2016_paper_26.pdf">
+	      The CLARIN Language Resource Switchboard.</a> CLARIN 2016 Annual Conference,
+	      Aix-en-Provence, France, 2016.</li>
+              <li>Claus Zinn, Marie Hinrichs, Emanuel Dima, Dieter van
+              Uytvanck. <a href="https://office.clarin.eu/v/CE-2015-0684-LR_switchboard_spec.pdf">The
+		  Switchboard specification </a>(Milestone 2.2 of the CLARIN-PLUS project).</li>
+              <li>Claus
+              Zinn. <a href="https://office.clarin.eu/v/CE-2016-0881-CLARINPLUS-D2_5.pdf">D2.5 LR
+              Switchboard (software)</a> Deliverable in the CLARIN-PLUS project.</li>
             </ul>
 	    
             <h3>Credits</h3>
@@ -76,10 +85,11 @@ class AboutHelpText extends React.Component {
               <li>Maarten van Gompel (integration of the CLAM Webservices)</li>
               <li>Bart Jongejan (integration of the CLARIN-DK Tool Box)</li>
 	      <li>Wojtek Rauk (integration of MorphoDoTa, WebSty, Morfeusz, Liner2 etc.)</li>		
-	      <li>Rafael Jaworski (integration of Concraft, Nerf, Spejd etc)</li>		
+	      <li>Rafael Jaworski (integration of Concraft, Nerf, Spejd etc.)</li>		
 	      <li>Josef Misutka and Pavel Stranak (UDPipe)</li>		
-              <li>Twan Goosen (integration of the VLO with the switchboard) </li>		
-              <li>Dieter Van Uytvanck (feedback on specification and usability)</li>
+              <li>Twan Goosen (switchboard integration in the VLO, Use Cases) </li>
+	      <li>André Moreira (Use Cases, Dockerization)</li>
+              <li>Dieter Van Uytvanck (feedback on specification and usability, Use Cases)</li>
               <li><em>Your Name Here</em> (integration of <em>your</em> tool)</li>
             </ul>
             <p>
@@ -103,12 +113,13 @@ class AboutHelpText extends React.Component {
             <h3>Technology used</h3>
             <ul>
               <li><a href="https://facebook.github.io/react/">ReactJS</a> (version 15.6.1)</li>
-	      <li><a href="https://nodejs.org/en/">Nodejs</a> (version 8.11.1)</li>
+	      <li><a href="https://nodejs.org/en/">Nodejs</a> (version 8.11.3)</li>
 	      <li><a href="https://www.npmjs.com">Javascript package manager npm</a> (version 5.6.0)</li>
 	      <li><a href="http://alt.js.org/docs/components/altContainer/">altjs</a> (version 0.18.6)</li>
 	      <li><a href="https://webpack.github.io">Webpack Javascript module bundler</a>(version 4.12.0)</li>
 	      <li><a href="http://www.json.org">JSON (JavaScript Object Notation)</a></li>
               <li><a href="http://tika.apache.org/">Apache Tika (tika-server-1.16.jar)</a></li>
+	      <li><a href="https://www.docker.com/">Docker</a> (for deployment)</li>
             </ul>
 	    
 	    <p>
@@ -118,7 +129,7 @@ class AboutHelpText extends React.Component {
           </div>
 
 	  <p>
-	    We value your feedback! For any questions or suggestions, please contact the <a href={ emailContactCommand }>Switchboard Team</a>. But please consult the user help beforehand (see top navigation bar).
+	    We value your feedback! For any questions or suggestions, please contact the <a href={ emailContactCommand }>Switchboard Team</a>. 
 	  </p>
         </ModalDialog>
 </ModalContainer>;